% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triagg-package.R
\docType{package}
\name{triagg-package}
\alias{triagg-package}
\alias{triagg}
\title{The 'triagg' package.}
\description{
An application to create consensus key population estimates (triangulator) and combine them into a single national estimate (aggregator)
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Carl Corcoran \email{ruu2@cdc.gov}

Authors:
\itemize{
  \item Ian Fellows \email{ian@fellstat.com}
}

}
